+++ components/qualaroo/sources/common/base.mjs
@@ -1,4 +1,5 @@
 import qualaroo from "../../qualaroo.app.mjs";
+import constants from "../../common/constants.mjs";
 import { DEFAULT_POLLING_SOURCE_TIMER_INTERVAL } from "@pipedream/platform";
 
 export default {
@@ -19,5 +20,38 @@ export default {
     setOffset(offset) {
       this.db.set("offset", offset);
     },
+  },
+  async run() {
+    let offset = this.getOffset();
+    const limit = constants.MAX_LIMIT;
+    const data = [];
+
+    const {
+      fn,
+      opts,
+    } = this.getListingFunctionOpts();
+
+    while (true) {
+      const response = await fn.call(this, {
+        ...opts,
+        params: {
+          ...opts.params,
+          limit,
+          offset,
+        },
+      });
+
+      if (response.length === 0) {
+        break;
+      }
+
+      offset += limit;
+      this.setOffset(offset);
+      data.push(...response);
+    }
+
+    for (const event of data) {
+      this.$emit(event, this.getMeta(event));
+    }
   },
 };
