+++ components/qualaroo/sources/survey-response-received/survey-response-received.mjs
@@ -1,5 +1,4 @@
 import base from "../common/base.mjs";
-import constants from "../../common/constants.mjs";
 
 export default {
   ...base,
@@ -18,34 +17,22 @@ export default {
       ],
     },
   },
-  async run() {
-    const data = [];
-    const limit = constants.MAX_LIMIT;
-    let offset = this.getOffset();
-
-    while (true) {
-      const response = await this.qualaroo.listSurveyesponses({
-        params: {
-          limit,
-          offset,
+  methods: {
+    ...base.methods,
+    getListingFunctionOpts() {
+      return {
+        fn: this.qualaroo.listSurveyResponses,
+        opts: {
+          survey: this.survey,
         },
-      });
-
-      if (response.length === 0) {
-        break;
-      }
-
-      offset += limit;
-      this.setOffset(offset);
-      data.push(...response);
-    }
-
-    for (const surveyResponse of data) {
-      this.$emit(surveyResponse,  {
+      };
+    },
+    getMeta(surveyResponse) {
+      return {
         id: surveyResponse.id,
         summary: `New response received for: ${surveyResponse.nudge_name}`,
         ts: new Date(surveyResponse.time),
-      });
-    }
+      };
+    },
   },
 };
